able to load all the cubemap textures, also added necessary resources for cubemap drowing

+++ application/assets/shaders/vulkan/axislines.vert
@@ -23,7 +23,7 @@ layout (set = 0, binding = 1) uniform Time
     double timeSinceStartup;
 } TIME;
 
-layout(location = 0) in vec2 inPosition;
+layout(location = 0) in ivec2 inPosition;
 layout(location = 0) out vec3 o_RayWorld;
 
 void main()
